chore: increase gas price in faucet.py
To avoid insufficient gas when the system is set up initially, increase
the gas price in faucet.py to 1000000000. That value comes from the
protocol_params.go file in go-ethereum where the `InitialBaseFee` for
EIP-1559 blocks is set to 1000000000.

Add a script that funds all tests addresses through faucet.py .